exclude comments from the word wrapping handling
instead of joining consecutive lines together, like it is done for
all other lines types, comments are directly discarded in the
preprocessing step and no joining is performed. the reason is that
the current word-wrapping behavior is common for markup (e.g. html,
latex) but definitely not common for line comments.